# docs/openapi-apifox.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "云舒 · 后端服务",
    "description": "AI 心理健康助手「云舒」后端服务。由 Java 版迁移而来，**接口契约与原版逐字对齐**。\n\n## 统一响应格式\n\n除 SSE 接口外，所有接口都返回同一个信封：\n\n```json\n{ \"code\": \"200\", \"msg\": \"success\", \"data\": {} }\n```\n\n- `code` 是**字符串**，`\"200\"` 表示成功，其余为业务错误码\n- **业务失败通常仍是 HTTP 200**，必须判 `code` 而不是 HTTP 状态码\n- 少数情况 HTTP 与 code 一起变：未登录 401、参数校验失败 400、越权 400\n\n## 鉴权\n\n登录 `/api/user/login` 拿到 `data.token`，之后所有需鉴权接口带：\n\n```\nAuthorization: Bearer <token>\n```\n\n**在 Apifox 里只需设置一次**：登录接口 → 后置操作 → 提取变量 `data.token`，\n然后把它绑到全局的 Bearer 认证即可（详见项目 README）。\n\n需鉴权的接口占 35/39。**无需鉴权的只有 4 个**：登录、注册、刷新令牌、登出。\n\n## 权限\n\n`userType`：`1` = 普通用户，`2` = 管理员。带 👑 的接口需要管理员，\n普通用户调用返回 **HTTP 400 + code `A0300`**（不是 403，这是原版的行为）。\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/knowledge/article/page": {
      "get": {
        "tags": [
          "知识库"
        ],
        "summary": "文章分页（管理端列表）",
        "operationId": "article_page_api_knowledge_article_page_get",
        "parameters": [
          {
            "name": "currentPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Currentpage"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoryid"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_dict_"
                },
                "example": {
                  "code": "200",
                  "msg": "操作成功",
                  "data": {
                    "records": [
                      {
                        "id": "db6467ec-f0e4-465f-96af-e22263176652",
                        "categoryId": 1,
                        "categoryName": "心理健康基础",
                        "title": "缓解考试焦虑的实用方法",
                        "summary": null,
                        "content": "考试焦虑是大学生常见的情绪反应。研究表明，考前进行深呼吸练习可以有效降低生理唤醒水平。具体做法：吸气四秒、屏息七秒、呼气八秒，重复四轮。此外，把复习计划拆成小块、保证睡眠、适度运动，都能显著缓解考试焦虑。如果焦虑持续两周以上并影响正常生活，建议前往学校心理健康教育中心寻求专业帮助。",
                        "coverImage": null,
                        "tags": null,
                        "authorName": null,
                        "readCount": 9,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2026-10-01T15:38:08",
                        "createdAt": "2026-10-01T15:38:08",
                        "updatedAt": "2026-10-02T14:46:19"
                      },
                      {
                        "id": "5a31f362-2b97-4858-8c5f-33168952f0f3",
                        "categoryId": 1,
                        "categoryName": "心理健康基础",
                        "title": "缓解考试焦虑的实用方法",
                        "summary": null,
                        "content": "考试焦虑是大学生常见的情绪反应。研究表明，考前进行深呼吸练习可以有效降低生理唤醒水平。具体做法：吸气四秒、屏息七秒、呼气八秒，重复四轮。此外，把复习计划拆成小块、保证睡眠、适度运动，都能显著缓解考试焦虑。如果焦虑持续两周以上并影响正常生活，建议前往学校心理健康教育中心寻求专业帮助。",
                        "coverImage": null,
                        "tags": null,
                        "authorName": null,
                        "readCount": 0,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2026-10-01T15:37:16",
                        "createdAt": "2026-10-01T15:37:16",
                        "updatedAt": "2026-10-01T15:37:16"
                      },
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440007",
                        "categoryId": 3,
                        "categoryName": "压力缓解",
                        "title": "学生心理压力应对策略",
                        "summary": "帮助学生识别和应对学习生活中的各种心理压力。",
                        "content": "<p>学生时期面临着学业、人际、未来规划等多重压力。掌握有效的应对策略对学生心理健康至关重要。</p><h3>常见压力源</h3><ul><li>学业成绩压力</li><li>同伴关系</li><li>家庭期望</li><li>未来规划焦虑</li></ul><h3>应对策略</h3><ol><li>合理规划学习时间</li><li>培养兴趣爱好</li><li>建立支持网络</li><li>学会放松技巧</li><li>寻求专业帮助</li></ol><h3>预防措施</h3><p>建立健康的学习习惯和生活方式。</p>",
                        "coverImage": "",
                        "tags": "学生心理,学业压力,青少年心理",
                        "authorName": "系统管理员",
                        "readCount": 37,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2025-09-07T08:30:00",
                        "createdAt": "2025-09-07T08:30:00",
                        "updatedAt": "2025-11-10T19:07:53"
                      },
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440006",
                        "categoryId": 2,
                        "categoryName": "情绪管理",
                        "title": "正念练习入门指南",
                        "summary": "正念练习是现代心理学推荐的有效减压和情绪调节方法。",
                        "content": "<p>正念是一种专注于当下体验的练习方法，能够帮助我们更好地管理情绪和压力。</p><h3>什么是正念</h3><p>正念是有意识地、不带评判地关注当下时刻的能力。</p><h3>正念的益处</h3><ul><li>减少焦虑和抑郁</li><li>提高专注力</li><li>改善情绪调节</li><li>增强自我觉察</li></ul><h3>基础正念练习</h3><ol><li>呼吸观察</li><li>身体扫描</li><li>行走冥想</li><li>日常正念</li></ol><h3>练习建议</h3><p>从每天5-10分钟开始，逐渐增加练习时间。</p>",
                        "coverImage": "",
                        "tags": "正念,冥想,减压,专注力",
                        "authorName": "系统管理员",
                        "readCount": 45,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2025-09-06T13:10:00",
                        "createdAt": "2025-09-06T13:10:00",
                        "updatedAt": "2025-09-06T13:10:00"
                      },
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440005",
                        "categoryId": 1,
                        "categoryName": "心理健康基础",
                        "title": "睡眠质量与心理健康",
                        "summary": "优质睡眠是维护心理健康的重要基础。",
                        "content": "<p>睡眠质量与心理健康密切相关。充足的睡眠不仅能恢复身体机能，还对情绪调节和认知功能有重要影响。</p><h3>睡眠对心理健康的影响</h3><ul><li>情绪稳定性</li><li>压力应对能力</li><li>记忆和学习能力</li><li>注意力集中</li></ul><h3>改善睡眠质量的方法</h3><ol><li>建立规律作息</li><li>创造良好睡眠环境</li><li>睡前放松活动</li><li>避免刺激性物质</li><li>适度运动</li></ol>",
                        "coverImage": "",
                        "tags": "睡眠健康,心理健康,生活习惯",
                        "authorName": "系统管理员",
                        "readCount": 27,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2025-09-05T11:45:00",
                        "createdAt": "2025-09-05T11:45:00",
                        "updatedAt": "2025-09-05T11:45:00"
                      },
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440004",
                        "categoryId": 4,
                        "categoryName": "人际关系",
                        "title": "建立健康的人际关系",
                        "summary": "良好的人际关系是心理健康和生活幸福的基石。",
                        "content": "<p>人际关系质量直接影响我们的心理健康和生活满意度。本文将探讨如何建立和维护健康的人际关系。</p><h3>健康关系的特征</h3><ul><li>相互尊重和信任</li><li>有效沟通</li><li>情感支持</li><li>个人边界清晰</li></ul><h3>沟通技巧</h3><ol><li>积极倾听</li><li>表达真实感受</li><li>非暴力沟通</li><li>冲突解决</li></ol><h3>维护关系的方法</h3><p>定期联系、表达感激、共同成长、处理分歧。</p>",
                        "coverImage": "",
                        "tags": "人际关系,沟通技巧,社交能力",
                        "authorName": "系统管理员",
                        "readCount": 18,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2025-09-04T16:20:00",
                        "createdAt": "2025-09-04T16:20:00",
                        "updatedAt": "2025-09-04T16:20:00"
                      },
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440003",
                        "categoryId": 3,
                        "categoryName": "压力缓解",
                        "title": "职场压力管理指南",
                        "summary": "在快节奏的工作环境中，学会有效管理压力是职业成功的关键。",
                        "content": "<p>职场压力是现代工作生活中不可避免的挑战。本指南将帮助您建立有效的压力管理体系。</p><h3>识别压力源</h3><ul><li>工作量过大</li><li>人际关系紧张</li><li>职业发展焦虑</li><li>工作与生活平衡困难</li></ul><h3>压力管理策略</h3><ol><li>时间管理技巧</li><li>设定合理目标</li><li>建立支持网络</li><li>培养兴趣爱好</li><li>定期休息和放松</li></ol><h3>长期压力预防</h3><p>建立健康的工作习惯和生活方式，预防压力累积。</p>",
                        "coverImage": "",
                        "tags": "职场压力,时间管理,工作生活平衡",
                        "authorName": "系统管理员",
                        "readCount": 31,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2025-09-03T09:15:00",
                        "createdAt": "2025-09-03T09:15:00",
                        "updatedAt": "2025-09-03T09:15:00"
                      },
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440002",
                        "categoryId": 2,
                        "categoryName": "情绪管理",
                        "title": "情绪调节的五个有效策略",
                        "summary": "掌握情绪调节技巧，让生活更加平衡和谐。",
                        "content": "<p>情绪调节是心理健康的重要组成部分。以下是五个经过科学验证的情绪调节策略：</p><h3>1. 认知重构</h3><p>学会从不同角度看待问题，挑战负面思维模式。</p><h3>2. 情绪标记</h3><p>准确识别和命名自己的情绪状态。</p><h3>3. 渐进式肌肉放松</h3><p>通过身体放松来缓解情绪紧张。</p><h3>4. 表达性写作</h3><p>通过写作来处理复杂的情绪体验。</p><h3>5. 社交支持</h3><p>寻求朋友和家人的理解与支持。</p>",
                        "coverImage": "",
                        "tags": "情绪调节,心理技巧,自我管理",
                        "authorName": "系统管理员",
                        "readCount": 23,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2025-09-02T14:30:00",
                        "createdAt": "2025-09-02T14:30:00",
                        "updatedAt": "2025-09-02T14:30:00"
                      },
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440001",
                        "categoryId": 1,
                        "categoryName": "心理健康基础",
                        "title": "如何识别和管理焦虑情绪",
                        "summary": "焦虑是现代生活中常见的情绪反应，学会识别和管理焦虑对心理健康至关重要。",
                        "content": "<p>焦虑是一种正常的情绪反应，但过度的焦虑可能会影响我们的日常生活。本文将介绍如何识别焦虑的症状，以及一些有效的管理方法。</p><h3>识别焦虑的症状</h3><ul><li>心跳加速、出汗</li><li>呼吸急促</li><li>思维混乱</li><li>难以集中注意力</li></ul><h3>管理焦虑的方法</h3><ol><li>深呼吸练习</li><li>正念冥想</li><li>适度运动</li><li>寻求专业帮助</li></ol>",
                        "coverImage": "",
                        "tags": "焦虑,情绪管理,心理健康",
                        "authorName": "系统管理员",
                        "readCount": 16,
                        "status": 1,
                        "statusText": "已发布",
                        "isFavorited": false,
                        "favoriteCount": 0,
                        "publishedAt": "2025-09-01T10:00:00",
                        "createdAt": "2025-09-01T10:00:00",
                        "updatedAt": "2025-09-01T10:00:00"
                      }
                    ],
                    "total": 9
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/article": {
      "get": {
        "tags": [
          "知识库"
        ],
        "summary": "文章分页（用户端列表）",
        "operationId": "article_list_api_knowledge_article_get",
        "parameters": [
          {
            "name": "currentPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Currentpage"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoryid"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "新增文章",
        "operationId": "article_create_api_knowledge_article_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_ArticleRow_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/article/{article_id}": {
      "get": {
        "tags": [
          "知识库"
        ],
        "summary": "文章详情",
        "description": "文章不存在时返回 code=200 + data=null（**不是 404**，原版如此）。",
        "operationId": "article_detail_api_knowledge_article__article_id__get",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Union_ArticleRow__NoneType__"
                },
                "example": {
                  "code": "200",
                  "msg": "操作成功",
                  "data": {
                    "id": "db6467ec-f0e4-465f-96af-e22263176652",
                    "categoryId": 1,
                    "categoryName": "心理健康基础",
                    "title": "缓解考试焦虑的实用方法",
                    "summary": null,
                    "content": "考试焦虑是大学生常见的情绪反应。研究表明，考前进行深呼吸练习可以有效降低生理唤醒水平。具体做法：吸气四秒、屏息七秒、呼气八秒，重复四轮。此外，把复习计划拆成小块、保证睡眠、适度运动，都能显著缓解考试焦虑。如果焦虑持续两周以上并影响正常生活，建议前往学校心理健康教育中心寻求专业帮助。",
                    "coverImage": null,
                    "tags": null,
                    "authorName": null,
                    "readCount": 10,
                    "status": 1,
                    "statusText": "已发布",
                    "isFavorited": false,
                    "favoriteCount": 0,
                    "publishedAt": "2026-10-01T15:38:08",
                    "createdAt": "2026-10-01T15:38:08",
                    "updatedAt": "2026-10-02T14:47:33"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "知识库"
        ],
        "summary": "更新文章",
        "operationId": "article_update_api_knowledge_article__article_id__put",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Union_ArticleRow__NoneType__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "知识库"
        ],
        "summary": "删除文章",
        "operationId": "article_delete_api_knowledge_article__article_id__delete",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/article/{article_id}/status": {
      "put": {
        "tags": [
          "知识库"
        ],
        "summary": "修改发布状态",
        "operationId": "article_update_status_api_knowledge_article__article_id__status_put",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleIn": {
        "properties": {
          "categoryId": {
            "type": "integer",
            "title": "Categoryid"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "coverImage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverimage"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "status": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  0,
                  1
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "categoryId",
          "title",
          "content"
        ],
        "title": "ArticleIn"
      },
...
      "ArticleStatusIn": {
        "properties": {
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ArticleStatusIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Result_ArticleRow_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "业务码，字符串类型",
            "default": "200"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "提示信息",
            "default": "操作成功"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArticleRow"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          }
        },
        "type": "object",
        "title": "Result[ArticleRow]"
      },
...
      "Result_NoneType_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "业务码，字符串类型",
            "default": "200"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "提示信息",
            "default": "操作成功"
          },
          "data": {
            "type": "null",
            "title": "Data",
            "description": "业务数据"
          }
        },
        "type": "object",
        "title": "Result[NoneType]"
      },
...
      "Result_Union_ArticleRow__NoneType__": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "业务码，字符串类型",
            "default": "200"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "提示信息",
            "default": "操作成功"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArticleRow"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          }
        },
        "type": "object",
        "title": "Result[Union[ArticleRow, NoneType]]"
      },
...
      "Result_dict_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "业务码，字符串类型",
            "default": "200"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "提示信息",
            "default": "操作成功"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "业务数据"
          }
        },
        "type": "object",
        "title": "Result[dict]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 data.token。在 Apifox 里**只填 token 本身**，会自动补 `Bearer ` 前缀。也可填 `{{token}}` 引用全局变量。"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1236",
      "description": "本地直连后端（推荐）"
    },
    {
      "url": "http://127.0.0.1:5173",
      "description": "经前端 Vite 代理（需前端在运行）"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}